Update tinder_bot.py
An error occurred when I got a match. Because I use the right key to swipe right, if a match popup appears, I would never be able to close out of it and continue because the Key.right doesn't return any error. To fix this, my finite state machine first looked for a popup and then continued swiping.